+++ DepRegAttributes.ExampleLibrary/RegisteredClasses/TaggedServices/TaggedWithType.cs
@@ -1,0 +1,12 @@
+﻿namespace DepRegAttributes.ExampleLibrary.RegisteredClasses.TaggedServices;
+
+public interface ITaggedWithType
+{
+}
+
+[RegisterTransient(types: typeof(ITaggedWithType), Tag = typeof(TypeTag))]
+public class TaggedWithType : ITaggedWithType
+{
+}
+
+public class TypeTag { }